+++ unstructured/ingest/cli/cli.py
@@ -24,6 +24,7 @@ subcommands = [
     cli_cmds.wikipedia,
     cli_cmds.gdrive,
     cli_cmds.biomed,
+    cli_cmds.notion,
     cli_cmds.onedrive,
     cli_cmds.outlook,
     cli_cmds.local,