+++ unstructured/ingest/main.py
@@ -1,6 +1,11 @@
 #!/usr/bin/env python3
 from unstructured.ingest.cli.cli import get_cmd
 
-if __name__ == "__main__":
+
+def main():
     ingest_cmd = get_cmd()
     ingest_cmd()
+
+
+if __name__ == "__main__":
+    main()